[BE] TEST: positon api 분기 & section 포지션 변경 케이스 추가
