+++ server/test/project.section.api.test.js
@@ -169,195 +169,6 @@ describe('create section', () => {
   });
 });
 
-describe('update section task positions', () => {
-  it('update section task positions 일반', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[0].id;
-    const expectedSectionId = seeder.sections[0].id;
-    const orderedTasks = seeder.tasks
-      .filter(task => task.sectionId === expectedSectionId)
-      .map(task => {
-        const { id, ...others } = task;
-        return id;
-      })
-      .sort((task1, task2) => task1.id - task2.id);
-
-    const requestBody = { orderedTasks };
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(status.SUCCESS.CODE);
-    expect(res.body.message).toBe(status.SUCCESS.MSG);
-    done();
-  });
-  it('taskId가 잘못된 경우 sectino task position 수정', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[0].id;
-    const expectedSectionId = seeder.sections[0].id;
-    const requestBody = { orderedTasks: ['hi', seeder.tasks[1].id, seeder.tasks[0].id] };
-    const expectedError = customError.INVALID_INPUT_ERROR();
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('projectID가 잘못된 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = 'invalidId';
-    const expectedSectionId = seeder.sections[0].id;
-    const requestBody = {
-      orderedTasks: [seeder.tasks[2].id, seeder.tasks[1].id, seeder.tasks[0].id],
-    };
-    const expectedError = customError.INVALID_INPUT_ERROR();
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('sectionId 잘못된 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[0].id;
-    const expectedSectionId = 'invalidId';
-    const requestBody = {
-      orderedTasks: [seeder.tasks[2].id, seeder.tasks[1].id, seeder.tasks[0].id],
-    };
-    const expectedError = customError.INVALID_INPUT_ERROR();
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('자신의 projectId가 아닌 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[2].id;
-    const expectedSectionId = seeder.sections[0].id;
-    const requestBody = {
-      orderedTasks: [seeder.tasks[2].id, seeder.tasks[1].id, seeder.tasks[0].id],
-    };
-    const expectedError = customError.FORBIDDEN_ERROR();
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('존재하지 않는 project인 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.sections[2].id;
-    const expectedSectionId = seeder.sections[0].id;
-    const requestBody = {
-      orderedTasks: [seeder.tasks[2].id, seeder.tasks[1].id, seeder.tasks[0].id],
-    };
-    const expectedError = customError.NOT_FOUND_ERROR('project');
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('존재하지 않는 section 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[0].id;
-    const expectedSectionId = seeder.sections[1].id;
-    const requestBody = {
-      orderedTasks: [seeder.tasks[2].id, seeder.tasks[1].id, seeder.tasks[0].id],
-    };
-    const expectedError = customError.NOT_FOUND_ERROR('section');
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-  it('존재하지 않는 taskId가 포함된 경우', async done => {
-    // given
-    const expectedUser = seeder.users[0];
-    const expectedProjectId = seeder.projects[0].id;
-    const expectedSectionId = seeder.sections[0].id;
-    const orderedTasks = seeder.tasks
-      .filter(task => task.sectionId === expectedSectionId)
-      .map(task => {
-        const { id, ...others } = task;
-        return id;
-      })
-      .sort((task1, task2) => task1.id - task2.id);
-    orderedTasks[0] = seeder.sections[2].id;
-    const requestBody = { orderedTasks };
-    const expectedError = customError.WRONG_RELATION_ERROR(
-      'please check projectId, sectionId, tasks Id',
-    );
-
-    // when
-    const res = await request(app)
-      .patch(`/api/project/${expectedProjectId}/section/${expectedSectionId}/task/position`)
-      .set('Authorization', `Bearer ${createJWT(expectedUser)}`)
-      .send(requestBody);
-
-    // then
-    expect(res.status).toBe(expectedError.status);
-    expect(res.body.code).toBe(expectedError.code);
-    expect(res.body.message).toBe(expectedError.message);
-    done();
-  });
-});
-
 describe('update section', () => {
   it('update section 일반', async done => {
     // given
